% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-package.R, R/explain.R
\docType{package}
\name{explain}
\alias{explain}
\alias{explain-package}
\title{Local Interpretable Model-agnostic Explanations.}
\usage{
explain(classifier, onerow, variable_summaries, num_features = 3,
  num_samples = 100, explainer = MASS::lm.ridge,
  feature_selection = "auto", ...)
}
\arguments{
\item{classifier}{tundraContainer. Can be either a \code{tundraContainer} or
a \code{list} with a \code{predict} method attached.}

\item{onerow}{data.frame. A one-row dataframe around which we want to get a
local approximation.}

\item{variable_summaries}{list. A named list that defines means and standard
deviations for the variables that need to be perturbed.}
}
\value{
an explanation object.
}
\description{
Local Interpretable Model-agnostic Explanations.

Generate an explanation for a classifier's prediction.
}

